RAD-24 - Study.modality not type safe
