+++ api/src/main/resources/liquibase.xml
@@ -1,10 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<databaseChangeLog xmlns="http://www.liquibase.org/xml/ns/dbchangelog/1.9"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog/1.9
-                  http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-1.9.xsd">
-
+<databaseChangeLog xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:ext="http://www.liquibase.org/xml/ns/dbchangelog-ext"
+	xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-2.0.xsd
+    http://www.liquibase.org/xml/ns/dbchangelog-ext http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-ext.xsd">
 
 	<!-- See http://www.liquibase.org/manual/home#available_database_refactorings 
 		for a list of supported elements and attributes -->
@@ -39,5 +38,15 @@
 				defaultValue="-1"></column>
 			<column name="mwl_status" type="int" defaultValue="-1"></column>
 		</createTable>
+	</changeSet>
+	<changeSet id="radiology-5" author="teleivo">
+		<comment>Change modality column type to varchar(16) in radiology_study
+			table to represent DICOM Tag Modality (0008,0060) value representation CS (Code String) defined in DICOM PS3.5 2014b - Data Structures and Encoding</comment>
+		<modifyDataType tableName="radiology_study" columnName="modality"
+			newDataType="varchar(16)" />
+		<dropDefaultValue tableName="radiology_study"
+			columnName="modality" columnDataType="varchar(16)" />
+		<addNotNullConstraint tableName="radiology_study"
+			columnName="modality" columnDataType="varchar(16)" />
 	</changeSet>
 </databaseChangeLog>
